// src/api/types.ts
export interface Category {
    id: number;
    name: string;
}

export interface Question {
    category: string;
    type: string;
    difficulty: string;
    question: string;
    correct_answer: string;
    incorrect_answers: string[];
}

export interface CategoryWithCount extends Category {
    questionCount: number;
    checked: boolean;
}